requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
torch>=1.9.0
torchaudio>=0.9.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0
webrtcvad>=2.0.10

# Speaker recognition
sidekit>=1.3.0
speechbrain>=0.5.0

# Data handling
pandas>=1.3.0
h5py>=3.1.0
pyyaml>=5.4.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
click>=8.0.0
pathlib2>=2.3.0

# Development dependencies
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
